chore(.gitignore): update file paths for ignored CSS files
The file paths for the ignored CSS files in the `.gitignore` file have been updated to use the correct directory structure.

fix(styles): update h1 styles in _base.scss and main.css

The styles for the `h1` element have been updated in both `_base.scss` and `main.css` files. The font size has been changed to `1.8rem`, the font weight to `700`, and the letter spacing to `-0.25px`. This change improves the visual appearance of the `h1` heading.